requirements.txt
nltk>=3.9.1
python-telegram-bot>=22.6